feat: makes rod correction drawing more user-friendly
Rod corrections are now conducted by clicking (and releasing) at a start position, then moving to the end position and then clicking again. After the first click a line is drawn between that start position and the current cursor position in the image. This process can be aborted by right-clicking.