feat: enables viewing pictures from 2 different cameras/folders
A second folder of images can be selected in the 2nd tab. The rod position data file is used for images in both tabs. If the "Persistent View" mode is active, the software tries to show images from the same frame for both cameras. It notifies users on the statusbar, if there is no image available for this frame on one of the camera views.
A message is displayed in the statusbar, if no rod position data is available for the shown frame.