feat: re-enables logging and undo operations by refactoring its design
Logging operations is now done by individual loggers that are provided to objects that perform loggable operations. The loggers also handle requests for reverting of the logged operations. The loggers are maintained by a supervisor class that also handles the display of operations in the GUI.
This refactoring enables logging and reverting of operations independently for multiple camera views that were introduced previously.