feat: enables automatic detection/usage of prior corrections
During loading of data the automatically assigned output folder is checked for files that match the desired rod position data files. The user can choose to use those files instead of the chosen folder. This is specifically useful when some corrections were applied already and the user continues with their work later.