Refactored data and file operations
Data and file operations are factored out to individual library files in the backend folder to provide more separation between the graphical and functional elements (e.g. for improved testability).
The library numpy is factored out and replaced by the built-in math module.
Fixes a bug when creating a new rod that was breaking the program, which was caused by an insufficient index determination process.
Fixes a bug when saving cleaned data, that would only save the cleaned data on the second saving operation.