feat: inserts scaling value of rod positions
Adds a scaling property for the rod positions to the user adjustable settings. This is allowed to be changed in the range of 0.00 to 99.99. It adjusts the scaling factor that is applied to the rod positions gathered from the *.csv files for adaptation of the data to the image space.