Test fixtures and standardized user actions
A variety of reasonable inital GUI states are defined and provided as
fixtures usable for end-to-end tests of the RodTracker.
Additionally, actions commonly performed by users are replicated with
the qtbot to be used as building blocks for use-case-like end-to-end
tests. These incluce for example rod number and position switching,
un-/redoing, switching of frames and more.
Tests/assertions are now performed before/after each user action.
This shall facilitate easy usecase construction without the need for
complex considerations of the RodTracker's endstate, as the state is
examined incrementally.