Usecase tests
Tests using the basic user action functions for testing common tasks
performed by users utilizing the GUI elements. The state of the program
is tested for correctness before/after each of these atomic actions.
Tests are grouped into tests of the atomic user actions and into more
complex combinations of these actions that shall mimic typical work
flows.
In parametrized tests the whole function would break, if one test is
not succeeding. This is mitigated with a try-finally block to ensure
teardown code is run.