Refactoring: Making RodTracker extendable
The RodTracker application is refactored such that it can be extended
more easily. All functionlity specific to rod shaped particles has been
factored out and will be provided as said extension later.
Default versions of image and position data handlers are defined.
Logging of actions is refactored.
Settings have been refactored and simplified.
The layout of the main window has been changed slightly to allow for
more versatile extensions.
Base versions of image interaction and utility tabs are provided.
These are used internally as well as by extensions to add new
functionality to the UI.
New menus in the menubar can be inserted by extensions.