Add option to use an example model for detection
Adds a button to use the published example model used for rod detection.
This will download the model's *.pt file from Zenodo and store it in RodTracker's log directory. Subsequent use of this button will reference this file and only attempts to download it again if it is no longer present. (See #75 and #85).